build: splitting main and main common
