+++ mobile/library/common/engine_common.cc
@@ -6,10 +6,9 @@
 namespace Envoy {
 
 EngineCommon::EngineCommon(int argc, const char* const* argv)
-    : options_(argc, argv, &MainCommon::hotRestartVersion, spdlog::level::info),
+    : options_(argc, argv, &MainCommonBase::hotRestartVersion, spdlog::level::info),
       base_(options_, real_time_system_, default_listener_hooks_, prod_component_factory_,
-            std::make_unique<PlatformImpl>(), std::make_unique<Random::RandomGeneratorImpl>(),
-            nullptr) {
+            platform_impl_, std::make_unique<Random::RandomGeneratorImpl>(), nullptr) {
   // Disabling signal handling in the options makes it so that the server's event dispatcher _does
   // not_ listen for termination signals such as SIGTERM, SIGINT, etc
   // (https://github.com/envoyproxy/envoy/blob/048f4231310fbbead0cbe03d43ffb4307fff0517/source/server/server.cc#L519).
